refactor(core): extract config lib-resolution domain into its own module
Split the lib-resolution domain out of the ~5k-line config/mod.rs monolith
into a new crates/tsz-core/src/config/lib_resolution.rs, following the
earlier extends/inheritance slice.

Moved (behavior-preserving): lib resolution entry points
(resolve_lib_files*, resolve_default_lib_files*), lib-directory discovery
and caches, the embedded-lib fallback, on-disk lib map building, legacy
name aliasing/normalization, /// <reference lib> extraction, and the
default/core lib-name-for-target helpers.

Left in mod.rs as separate concerns: lib option validation
(validate_lib_values / find_lib_entry_offset and the VALID_LIB_* tables,
which belong to the validation domain) and checker_target_from_emitter
(a target conversion, not lib resolution).

The public config API path is preserved via pub use, so downstream
consumers (CLI, WASM, LSP) compile unchanged. config/mod.rs drops from
4981 to 4275 lines; lib_resolution.rs is 781 lines with a focused
unit-test module for its private helpers.

Refs #8280
